# specs/2-plan/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Physical AI Textbook API
  version: "1"
  description: >
    FastAPI backend for the Physical AI & Humanoid Robotics textbook.
    All endpoints live under /api/v1.
servers:
  - url: https://<render-service>.onrender.com
    description: Production (Render free tier)
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Better-Auth session token
  schemas:
...
    ErrorResponse:
      type: object
      required: [apiVersion, requestId, error]
      properties:
        apiVersion:
          type: string
          example: v1
        requestId:
          type: string
          example: req_chat_001
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - AUTH_REQUIRED
                - VALIDATION_ERROR
                - RATE_LIMITED
                - NOT_FOUND
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
...
paths:
...
  /api/v1/agents/{agentName}:
    post:
      operationId: invokeAgent
      summary: Invoke a subagent (quiz_generator, chapter_summarizer, prerequisite_mapper)
      description: >
        Authenticated endpoint for all three subagents.
        All responses are grounded in chapter content from the vector store.
        Rate limit: 20 req/min per authenticated user.
        p95 latency target ≤ 5s.
      security:
        - BearerAuth: []
      parameters:
        - name: agentName
          in: path
          required: true
          schema:
            type: string
            enum: [quiz_generator, chapter_summarizer, prerequisite_mapper]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: object
                  description: Agent-specific input
                  oneOf:
                    - title: quiz_generator input
                      properties:
                        chapterId:
                          type: string
                        difficulty:
                          type: string
                          enum: [easy, medium, hard]
                        count:
                          type: integer
                          minimum: 1
                          maximum: 10
                          default: 5
                    - title: chapter_summarizer input
                      properties:
                        chapterId:
                          type: string
                        level:
                          type: string
                          enum: [beginner, intermediate, advanced]
                    - title: prerequisite_mapper input
                      properties:
                        topic:
                          type: string
                        chapterId:
                          type: string
      responses:
        "200":
          description: Agent output
          content:
            application/json:
              schema:
                type: object
                required: [apiVersion, requestId, agentName, output]
                properties:
                  apiVersion:
                    type: string
                    example: v1
                  requestId:
                    type: string
                  agentName:
                    type: string
                  output:
                    type: object
                    description: Agent-specific output (see examples below)
              examples:
                quiz_generator:
                  value:
                    apiVersion: v1
                    requestId: req_agent_001
                    agentName: quiz_generator
                    output:
                      questions:
                        - question: What is VLA?
                          options: [A, B, C, D]
                          answer: B
                          explanation: VLA stands for Vision-Language-Action...
                chapter_summarizer:
                  value:
                    apiVersion: v1
                    requestId: req_agent_002
                    agentName: chapter_summarizer
                    output:
                      summary: This chapter covers...
                      keyPoints:
                        - ROS 2 uses DDS for communication
                        - Publishers and subscribers are decoupled
                        - QoS profiles control reliability
                      citations:
                        - chapterId: module-1-ros2-week1
                          sectionId: dds-overview
                          sourceUrl: /docs/module-1/ros2-week1#dds-overview
                prerequisite_mapper:
                  value:
                    apiVersion: v1
                    requestId: req_agent_003
                    agentName: prerequisite_mapper
                    output:
                      prerequisites:
                        - concept: ROS 2 Node Lifecycle
                          recommendedSection: /docs/module-1/ros2-week1#node-lifecycle
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Auth endpoints (Better-Auth passthrough) ─────────────────────────────
...